Move viewsets to views
The majority of this commit is renaming viewsets to views and the resulting refactor. The other change is removing action_buttons declaration so that the UIViewSet picks up the default. The outlier is CapircaPolicy which went from `action_buttons = None` to `action_buttons = []`
