+++ nautobot_firewall_models/urls.py
@@ -5,27 +5,27 @@ from django.urls import path
 from django.views.generic import RedirectView
 from nautobot.apps.urls import NautobotUIViewSetRouter
 
-from nautobot_firewall_models import viewsets
+from nautobot_firewall_models import views
 
 app_name = "nautobot_firewall_models"
 router = NautobotUIViewSetRouter()
-router.register("address-object", viewsets.AddressObjectUIViewSet)
-router.register("address-object-group", viewsets.AddressObjectGroupUIViewSet)
-router.register("application-object", viewsets.ApplicationObjectUIViewSet)
-router.register("application-object-group", viewsets.ApplicationObjectGroupUIViewSet)
-router.register("capirca-policy", viewsets.CapircaPolicyUIViewSet)
-router.register("capirca-policy-device", viewsets.CapircaPolicyDeviceUIViewSet, basename="capircapolicy_devicedetail")
-router.register("fqdn", viewsets.FQDNUIViewSet)
-router.register("ip-range", viewsets.IPRangeUIViewSet)
-router.register("nat-policy", viewsets.NATPolicyUIViewSet)
-router.register("nat-policy-rule", viewsets.NATPolicyRuleUIViewSet)
-router.register("policy", viewsets.PolicyUIViewSet)
-router.register("policy-rule", viewsets.PolicyRuleUIViewSet)
-router.register("service-object", viewsets.ServiceObjectUIViewSet)
-router.register("service-object-group", viewsets.ServiceObjectGroupUIViewSet)
-router.register("user-object", viewsets.UserObjectUIViewSet)
-router.register("user-object-group", viewsets.UserObjectGroupUIViewSet)
-router.register("zone", viewsets.ZoneUIViewSet)
+router.register("address-object", views.AddressObjectUIViewSet)
+router.register("address-object-group", views.AddressObjectGroupUIViewSet)
+router.register("application-object", views.ApplicationObjectUIViewSet)
+router.register("application-object-group", views.ApplicationObjectGroupUIViewSet)
+router.register("capirca-policy", views.CapircaPolicyUIViewSet)
+router.register("capirca-policy-device", views.CapircaPolicyDeviceUIViewSet, basename="capircapolicy_devicedetail")
+router.register("fqdn", views.FQDNUIViewSet)
+router.register("ip-range", views.IPRangeUIViewSet)
+router.register("nat-policy", views.NATPolicyUIViewSet)
+router.register("nat-policy-rule", views.NATPolicyRuleUIViewSet)
+router.register("policy", views.PolicyUIViewSet)
+router.register("policy-rule", views.PolicyRuleUIViewSet)
+router.register("service-object", views.ServiceObjectUIViewSet)
+router.register("service-object-group", views.ServiceObjectGroupUIViewSet)
+router.register("user-object", views.UserObjectUIViewSet)
+router.register("user-object-group", views.UserObjectGroupUIViewSet)
+router.register("zone", views.ZoneUIViewSet)
 
 urlpatterns = [
     path("docs/", RedirectView.as_view(url=static("nautobot_firewall_models/docs/index.html")), name="docs"),
