+++ nautobot_firewall_models/views/__init__.py
@@ -1,22 +1,22 @@
 """UI Viewsets."""
 
-from nautobot_firewall_models.viewsets.address import (
+from nautobot_firewall_models.views.address import (
     AddressObjectGroupUIViewSet,
     AddressObjectUIViewSet,
     FQDNUIViewSet,
     IPRangeUIViewSet,
 )
-from nautobot_firewall_models.viewsets.capirca_policy import CapircaPolicyDeviceUIViewSet, CapircaPolicyUIViewSet
-from nautobot_firewall_models.viewsets.nat_policy import NATPolicyRuleUIViewSet, NATPolicyUIViewSet
-from nautobot_firewall_models.viewsets.security_policy import PolicyRuleUIViewSet, PolicyUIViewSet
-from nautobot_firewall_models.viewsets.service import (
+from nautobot_firewall_models.views.capirca_policy import CapircaPolicyDeviceUIViewSet, CapircaPolicyUIViewSet
+from nautobot_firewall_models.views.nat_policy import NATPolicyRuleUIViewSet, NATPolicyUIViewSet
+from nautobot_firewall_models.views.security_policy import PolicyRuleUIViewSet, PolicyUIViewSet
+from nautobot_firewall_models.views.service import (
     ApplicationObjectGroupUIViewSet,
     ApplicationObjectUIViewSet,
     ServiceObjectGroupUIViewSet,
     ServiceObjectUIViewSet,
 )
-from nautobot_firewall_models.viewsets.user import UserObjectGroupUIViewSet, UserObjectUIViewSet
-from nautobot_firewall_models.viewsets.zone import ZoneUIViewSet
+from nautobot_firewall_models.views.user import UserObjectGroupUIViewSet, UserObjectUIViewSet
+from nautobot_firewall_models.views.zone import ZoneUIViewSet
 
 __all__ = (
     "AddressObjectUIViewSet",
